+++ CotEditor/Sources/KeyBindingItem.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2020 1024jp
+//  © 2016-2022 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -30,13 +30,15 @@ final class KeyBindingItem {
     
     // MARK: Public Properties
     
+    let name: String
     let action: Selector
     var shortcut: Shortcut?
     let defaultShortcut: Shortcut
     
     
-    init(action: Selector, shortcut: Shortcut?, defaultShortcut: Shortcut) {
+    init(name: String, action: Selector, shortcut: Shortcut?, defaultShortcut: Shortcut) {
         
+        self.name = name
         self.action = action
         self.shortcut = shortcut
         self.defaultShortcut = defaultShortcut